I realized I should have uploaded this as a patch set over my other cr, but I couldn't figure out how after merging with Kashi's code.
Added the abstract class FileSystem, which RealFileSystem and FakeFileSystem are derived from. FakeFileSystem is what FileSystem previously was, and RealFileSystem uses the <filesystem> functions to interact with the actual filesystem. The calls to filesystem have been abstracted away using the class FileSystem in EntityManager::insert, and ::get_entity.

To Do:
Make it so that PUT and DELETE uses the FileSystem object
Write unit tests for CrudRequestHandler / EntityManager PUT and DELETE
Integration tests

Change-Id: I4ad537a334f1176ac900f7cec6412e6ddb31a264