Add support for the GET /chess endpoint
What:
The GET /chess endpoint returns the HTML file of the home page.

Why:
When the user first enters the website, they should be able to
generate a new game via a button on the home.html file.

How:
Given a GET request with /chess URI, respond with the content of
the home.html file.

Testing:
Write unit tests for when the home.html file exists and when
it doesn't.

Change-Id: Iff73239330101a5833240d323061c657204be506